server/requirements.txt
# web framework and server
fastapi>=0.115.0
uvicorn[standard]>=0.31.0

# db & migration
sqlalchemy>=2.0.30
alembic>=1.13.1
psycopg2-binary>=2.9.9

# validation
pydantic>=2.9.0
pydantic-settings>=2.4.0

# Security & Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
python-multipart>=0.0.9

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0